+++ Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/targethandlers/affectscope/SquarePB.java
@@ -45,7 +45,7 @@ public class SquarePB implements IAffectScopeHandler {
 	public void forEachAffected(Creature activeChar, WorldObject target, Skill skill, Consumer<? super WorldObject> action) {
 		final int squareLength = skill.getFanRadius();
 		final int squareWidth = skill.getFanAngle();
-		final int radius = (int) Math.sqrt(squareLength * squareLength + squareWidth * squareWidth);
+		final int radius = (int) Math.sqrt((double) squareLength * squareLength + (double) squareWidth * squareWidth);
 
 		World.getInstance().forEachVisibleObjectInRange(activeChar, Creature.class, radius, action::accept, squareFilterOf(activeChar, skill, squareLength, squareWidth));
 	}
